refactor: Moved gamepad specific configs and mappings to own packages
